Improve support for weird links
